chore(lang): Update Hungarian translation
There are some changes that might seem strange, I'll try to describe why
I used them.

1. Flip coin -> Érme feldobása

We don't flip the coin, we throw it up. We "flip" a card, but "throw" a
coin.

2. Heads or Tails -> Fej vagy Írás

Historically, on coins we have 2 sides. One with a head on it and one
with the numerical value on it. Based on that we call it "Fej" (head)
and "Írás" (writing).

3. _From and _To

I tried to come up with a shorter version, but nothign really felt
rights. With "Legkisebb" and "Legnagyobb", the UI is breaking, but the
other option would be very ugly using "Tól" and "Ig". It is not alien to
the language to use that, but it's coming mostly from the early and mid 90s
computer boom when everyone tried to localize everything.

4. Head -> Fejek -> Fej

We don't say "heads" when we flip a coin, unless we do multiple ones at
the same time. It's singular.

5. Circled -> Karikás

Well, the raw translation would be "Karikázva" or "Bekarikázott". The
first one heavily implies there should another word behind it "Karikázva lett"
and without that, it's so painful to say that word alone. The second one
is just aweful I think.

Related to: #8